pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
# See https://www.python.org/dev/peps/pep-0621/
name = "aiida-fireball"
dynamic = ["version"]  # read from aiida_fireball/src/__init__.py
description = "AiiDA plugin for Fireball"
authors = [
    {name = "Valk Scripter", email = "valkscode@gmail.com"},]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Development Status :: 3 - Alpha",
    "Framework :: AiiDA"
]
keywords = ["aiida", "plugin"]
requires-python = ">=3.9"
dependencies = [
    "aiida-core>=2.5,<3",
    "voluptuous",
    'aiida-pseudo',
    'click',
    'importlib_resources',
    'jsonschema',
    'numpy',
    'pydantic',
    'packaging',
    'xmlschema',
    'ase',
    'pymc',
]

[project.optional-dependencies]
docs = [
    'sphinx~=6.2.1',
    'sphinx_rtd_theme',
    'sphinx-copybutton~=0.5.2',
    'sphinx-book-theme~=1.0.1',
    'sphinx-click~=4.4.0',
    'sphinx-design~=0.4.1',
    'sphinxcontrib-details-directive~=0.1.0',
    'sphinx-autoapi~=3.0.0',
    'myst-parser~=3.0.0',
    'linkify-it-py',
]
pre-commit = [
    'pre-commit',
    'pylint',
    'pylint-aiida',
    'toml',
    'hatch',
]
tests = [
    'pgtest',
    'pytest',
    'pytest-regressions',
    'pytest-cov',
]

[dependency-groups]
dev = [
  "aiida-fireball[docs,tests,pre-commit]",
]

[project.urls]
Source = "https://github.com/ValkScripter/aiida-fireball"

[project.entry-points."aiida.data"]

[project.entry-points."aiida.calculations"]
"fireball.structure.interpolate" = "aiida_fireball.calculations.structure.tools:interpolate_structures"
"fireball.structure.scale" = "aiida_fireball.calculations.structure.tools:scale_structure"
"fireball.fireball" = "aiida_fireball.calculations.fireball:FireballCalculation"

[project.entry-points."aiida.workflows"]
"fireball.relax"   = "aiida_fireball.workflows.relax:FireballRelaxChain"
"fireball.scf"     = "aiida_fireball.workflows.scf:FireballSCFChain"
"fireball.dos"     = "aiida_fireball.workflows.dos:FireballDOSChain"
"fireball.kpoints" = "aiida_fireball.workflows.kpoints:FireballKpointsChain"

[project.entry-points."aiida.parsers"]
"fireball.fireball" = "aiida_fireball.parsers.fireball:FireballParser"

[project.entry-points."aiida.cmdline.data"]

[tool.pytest.ini_options]
# Configuration for [pytest](https://docs.pytest.org)
python_files = "test_*.py example_*.py"
addopts = "--pdbcls=IPython.terminal.debugger:TerminalPdb"
filterwarnings = [
    "ignore::DeprecationWarning:aiida:",
    "ignore:Creating AiiDA configuration folder:",
    "ignore::DeprecationWarning:plumpy:",
    "ignore::DeprecationWarning:yaml:",
]

[tool.flynt]
line-length = 150
fail-on-change = true

[tool.isort]
force_sort_within_sections = true
include_trailing_comma = true
line_length = 150
multi_line_output = 3  # this configuration is compatible with yapf

[tool.coverage.run]
# Configuration of [coverage.py](https://coverage.readthedocs.io)
# reporting which lines of your plugin are covered by tests
source = ["src/aiida_fireball"]

[tool.ruff]
line-length = 150

[tool.ruff.lint]
ignore = [
    "ARG002",  # https://docs.astral.sh/ruff/rules/unused-method-argument/
    "ARG004",  # https://docs.astral.sh/ruff/rules/unused-static-method-argument/
    "B028",    # https://docs.astral.sh/ruff/rules/no-explicit-stacklevel/
    "BLE001",  # https://docs.astral.sh/ruff/rules/blind-except/
    "EM101",   # https://docs.astral.sh/ruff/rules/raw-string-in-exception/
    "EM102",   # https://docs.astral.sh/ruff/rules/f-string-in-exception/
    "EXE001",  # https://docs.astral.sh/ruff/rules/shebang-not-executable/
    "FA100",   # https://docs.astral.sh/ruff/rules/future-rewritable-type-annotation/
    "FBT001",  # https://docs.astral.sh/ruff/rules/boolean-type-hint-positional-argument/
    "FBT002",  # https://docs.astral.sh/ruff/rules/boolean-default-value-positional-argument/
    "G004",    # https://docs.astral.sh/ruff/rules/logging-f-string/
    "N805",    # https://docs.astral.sh/ruff/rules/invalid-first-argument-name-for-method/
    "PERF401", # https://docs.astral.sh/ruff/rules/manual-list-comprehension/
    "PLR2004", # https://docs.astral.sh/ruff/rules/magic-value-comparison/
    "RUF005",  # https://docs.astral.sh/ruff/rules/collection-literal-concatenation/
    "RUF012",  # https://docs.astral.sh/ruff/rules/mutable-class-default/
    "RET503",  # https://docs.astral.sh/ruff/rules/implicit-return/
    "S110",    # https://docs.astral.sh/ruff/rules/try-except-pass/
    "S318",    # https://docs.astral.sh/ruff/rules/suspicious-xml-mini-dom-usage/
    "S314",    # https://docs.astral.sh/ruff/rules/suspicious-xml-element-tree-usage/
    "TID252",  # https://docs.astral.sh/ruff/rules/relative-imports/
    "TRY004",  # https://docs.astral.sh/ruff/rules/type-check-without-type-error/
    "TRY003",  # https://docs.astral.sh/ruff/rules/raise-vanilla-args/
    "TRY301",  # https://docs.astral.sh/ruff/rules/raise-within-try/
]
select = [
#  'E',  # pydocstyle
#  'W',  # pydocstyle
  'F',  # pyflakes
#  'I',  # isort
#  'N',  # pep8-naming
#  'RUF'  # ruff
]

[tool.ruff.lint.per-file-ignores]
"tests/conftest.py" = [
    "PLC0415",  # https://docs.astral.sh/ruff/rules/import-outside-top-level/
]
"src/aiida_fireball/calculations/utils.py" = [
    "PLR0912",  # https://docs.astral.sh/ruff/rules/too-many-branches/
]


[tool.pylint.master]
load-plugins = ['pylint_aiida','pylint.extensions.no_self_use']

[tool.pylint.format]
max-line-length = 150

[tool.pylint.tool-check]
generated-members = 'self.exit_codes.*'

[tool.pylint.messages_control]
disable = [
    'duplicate-code',
    'fixme',
    'inconsistent-return-statements',
    'import-outside-toplevel',
    'locally-disabled',
    'logging-format-interpolation',
    'no-else-raise',
    'too-many-arguments',
    'too-many-ancestors',
    'too-many-branches',
    'too-many-locals',
    'use-dict-literal',
]

[tool.pylint.basic]
good-names = [
    'i',
    'j',
    'k',
    'StructureData',
    'SinglefileData',
]

[tool.yapf]
align_closing_bracket_with_visual_indent = true
based_on_style = 'google'
coalesce_brackets = true
column_limit = 150
dedent_closing_brackets = true
indent_dictionary_value = false
split_arguments_when_comma_terminated = true

## Hatch configurations

[tool.hatch.build.targets.wheel]
packages = ['src/aiida_fireball']

[tool.hatch.build.targets.sdist]
exclude = [
  'docs/',
  'tests/',
]

[tool.hatch.version]
path = 'src/aiida_fireball/__init__.py'

[tool.hatch.envs.default]
installer = 'uv'

[tool.hatch.envs.pre-commit]
features = ["pre-commit"]
scripts.install = 'pre-commit install'
scripts.run = 'pre-commit run {args}'

[tool.hatch.envs.hatch-test]
features = ["tests"]
randomize = false
parallel = false
run = "pytest {args}"

[[tool.hatch.envs.hatch-test.matrix]]
python = ["3.9", "3.10", "3.11", "3.12"]

[tool.hatch.envs.docs]
features = ["docs"]
scripts.clean = 'make -C docs clean'
scripts.build = 'make -C docs html'
scripts.view = 'make -C docs view'


[tool.hatch.envs.hatch-static-analysis]
dependencies = ["ruff==0.4.3"]

[tool.hatch.envs.hatch-static-analysis.scripts]
# Fixes are executed with `hatch fmt`.
# Checks are executed with `hatch fmt --check`.

format-check = "ruff format --check --config pyproject.toml {args:.}"
format-fix = "ruff format --config pyproject.toml {args:.}"
lint-check = "ruff check --config pyproject.toml {args:.}"
lint-fix = "ruff check --config pyproject.toml --fix --exit-non-zero-on-fix --show-fixes {args:.}"
